backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
google-generativeai
faiss-cpu
numpy
python-dotenv
tqdm
pydantic
PyPDF2